Sistematata gestione carte loy usabili in signola o più transazioni con parametro nell'ini

+++ PromoCalculator/BaseSystem.h
@@ -46,7 +46,7 @@ public:
     std::map <unsigned long long, Item> itemsMap ;
     std::map <unsigned long long, Barcodes> barcodesMap ;
     std::map <unsigned long long, Cart> cartsMap ;
-    std::map <unsigned long long, unsigned int> allLoyCardsMap ;
+    std::map <unsigned long long, unsigned long long> allLoyCardsMap ;
     
     typedef boost::shared_ptr<tcp::socket> socket_ptr;
 	std::map<unsigned long long, Cart>::iterator mainIterator ;
